% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{add_comment}
\alias{add_comment}
\title{Add comment}
\usage{
add_comment(team_name, post_num, body_md = NULL)
}
\arguments{
\item{team_name}{character}

\item{post_num}{integer}

\item{body_md}{character}
}
\description{
Add comment
}
\examples{
\dontrun{
add_comment("us-lab", post_num = 15, body_md = "Test comment")
}
}

