% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{create_post}
\alias{create_post}
\title{Create new post}
\usage{
create_post(team_name, post_name = NULL, body_md = NULL, ...)
}
\arguments{
\item{team_name}{character}

\item{post_name}{character. required}

\item{body_md}{character.}

\item{...}{Additional parameters}
}
\description{
Create new post
}
\examples{
\dontrun{
create_post("us-lab",
post_name = paste("sandbox", Sys.Date(), sep = "/"),
body_md = "Test",
tags = array("sandbox"),
wip = FALSE)
}
}

