% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{create_rmd_post}
\alias{create_rmd_post}
\title{Create new post from RMarkdown}
\usage{
create_rmd_post(team_name, file = NULL, post_name = NULL, ...)
}
\arguments{
\item{team_name}{character}

\item{file}{Input file (Rmd)}

\item{post_name}{character. required}

\item{...}{Additional parameters}
}
\description{
Create new post from RMarkdown
}
\examples{
\dontrun{
create_rmd_post("us-lab", file = "test.Rmd",
 post_name = "Post Rmarkdown file", tags = "sandbox")
}
}

