% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{delete_comment}
\alias{delete_comment}
\title{Delete comment}
\usage{
delete_comment(team_name, comment_id = NULL)
}
\arguments{
\item{team_name}{character}

\item{comment_id}{integer}
}
\description{
Delete comment
}
\examples{
\dontrun{
delete_comment("us-lab", comment_id = 79601)
}
}

