% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{delete_post}
\alias{delete_post}
\title{Delete post}
\usage{
delete_post(team_name, post_num)
}
\arguments{
\item{team_name}{character}

\item{post_num}{integer}
}
\description{
Delete post
}
\examples{
\dontrun{
delete_post("us-lab", 17)
}
}

