% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{description_post}
\alias{description_post}
\title{Get post description}
\usage{
description_post(team_name, post_num)
}
\arguments{
\item{team_name}{character.}

\item{post_num}{integer.}
}
\description{
Get post description
}
\examples{
description_post(team_name = "us-lab", post_num = 17)
}

