% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esaio_api.R
\name{esaio_api}
\alias{esaio_api}
\title{esa.io v1 wrapper function}
\usage{
esaio_api(verb = c("GET", "POST", "DELETE", "PATCH", "PUT"), path = NULL,
  body = NULL)
}
\arguments{
\item{verb}{character}

\item{path}{character}

\item{body}{list}
}
\description{
esa.io v1 wrapper function
}

