% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{modify_comment}
\alias{modify_comment}
\title{Modify comment}
\usage{
modify_comment(team_name, comment_id = NULL, body_md = NULL)
}
\arguments{
\item{team_name}{character}

\item{comment_id}{integer}

\item{body_md}{character}
}
\description{
Modify comment
}
\examples{
\dontrun{
modify_comment("us-lab",
 comment_id = 79598,
 body_md = "Change comment")
}
}

