% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{modify_post}
\alias{modify_post}
\title{Modify post}
\usage{
modify_post(team_name, post_num, post_name = NULL, body_md = NULL,
  append = TRUE, ...)
}
\arguments{
\item{team_name}{character}

\item{post_num}{integer}

\item{post_name}{character}

\item{body_md}{character}

\item{append}{logical}

\item{...}{Additional parameters}
}
\description{
Modify post
}
\examples{
\dontrun{
modify_post("us-lab", post_num = 17,
 post_name = "New Title", body_md = "Append mode", append = TRUE)
modify_post("us-lab", post_num = 17,
 body_md = "Overwrite mode", append = FALSE)
}
}

